+++ kikuchipy/signals/tests/test_ebsd.py
@@ -1291,24 +1291,91 @@ class TestEBSDDetectorProperty:
 
 
 class TestDictionaryIndexing:
-    def test_dictionary_indexing(self, dummy_signal):
-        """Scores are all 1.0 for a dictionary containing all patterns
+    def test_dictionary_indexing_doesnt_change_data(self, dummy_signal):
+        """Scores are all 1 for a dictionary containing all patterns
         from dummy_signal().
         """
         s_dict = kp.signals.EBSD(dummy_signal.data.reshape(-1, 3, 3))
         s_dict.axes_manager[0].name = "x"
         s_dict.xmap = CrystalMap.empty((9,))
-        xmap = dummy_signal.dictionary_indexing(s_dict)
+        dummy_signal2 = dummy_signal.deepcopy()
+        s_dict2 = s_dict.deepcopy()
+        xmap = dummy_signal2.dictionary_indexing(s_dict2, metric="ndp", rechunk=True)
 
+        assert isinstance(xmap, CrystalMap)
         assert np.allclose(xmap.scores[:, 0], 1)
 
-    def test_dictionary_indexing_warns(self, dummy_signal):
-        # TODO: Remove test after v0.4 is released
+        # Data is not affected by indexing method
+        assert np.allclose(dummy_signal.data, dummy_signal2.data)
+        assert np.allclose(s_dict.data, s_dict2.data)
+
+    def test_dictionary_indexing_signal_mask(self, dummy_signal):
+        """Passing a signal mask works, using 64-bit floats works,
+        rechunking of experimental patterns works.
+        """
         s_dict = kp.signals.EBSD(dummy_signal.data.reshape(-1, 3, 3))
         s_dict.axes_manager[0].name = "x"
-        s_dict.xmap = CrystalMap.empty((9,), step_sizes=(1,))
-        with pytest.warns(np.VisibleDeprecationWarning, match="Function "):
-            _ = dummy_signal.match_patterns(s_dict)
+        s_dict.xmap = CrystalMap.empty((9,))
+        signal_mask = np.array([[0, 0, 0], [0, 1, 0], [0, 0, 0]], dtype=bool)
+        xmap = dummy_signal.dictionary_indexing(
+            s_dict,
+            dtype=np.float64,
+            n_per_iteration=2,
+            signal_mask=signal_mask,
+            rechunk=True,
+        )
+        assert np.allclose(xmap.scores[:, 0], 1)
+
+    def test_dictionary_indexing_n_per_iteration_from_lazy(self, dummy_signal):
+        """Getting number of iterations from Dask array chunk works, and
+        NDP rechunking of experimental patterns works.
+        """
+        s_dict = kp.signals.EBSD(dummy_signal.data.reshape(-1, 3, 3))
+        s_dict.axes_manager[0].name = "x"
+        s_dict.xmap = CrystalMap.empty((9,))
+        s_dict_lazy = s_dict.as_lazy()
+        s_dict_lazy.xmap = s_dict.xmap
+        signal_mask = np.array([[0, 0, 0], [0, 1, 0], [0, 0, 0]], dtype=bool)
+        xmap = dummy_signal.dictionary_indexing(
+            s_dict_lazy, metric="ndp", signal_mask=signal_mask
+        )
+        assert np.allclose(xmap.scores[:, 0], 1)
+
+        # So that computing parts of the dictionary during indexing is
+        # covered, and NDP rechunking
+        xmap4 = dummy_signal.dictionary_indexing(
+            s_dict_lazy, metric="ndp", n_per_iteration=2, rechunk=True
+        )
+        assert np.allclose(xmap4.scores[:, 0], 1)
+
+    def test_dictionary_indexing_invalid_metric(self, dummy_signal):
+        s_dict = kp.signals.EBSD(dummy_signal.data.reshape(-1, 3, 3))
+        s_dict.axes_manager[0].name = "x"
+        s_dict.xmap = CrystalMap.empty((9,))
+        with pytest.raises(ValueError, match="'invalid' must be either of "):
+            _ = dummy_signal.dictionary_indexing(s_dict, metric="invalid")
+
+    def test_dictionary_indexing_invalid_signal_shapes(self, dummy_signal):
+        s_dict_data = dummy_signal.data[:, :, :2, :2].reshape((-1, 2, 2))
+        s_dict = kp.signals.EBSD(s_dict_data)
+        s_dict.axes_manager[0].name = "x"
+        s_dict.xmap = CrystalMap.empty((9,))
+        with pytest.raises(ValueError):
+            _ = dummy_signal.dictionary_indexing(s_dict)
+
+    def test_dictionary_indexing_invalid_dictionary(self, dummy_signal):
+        s_dict = kp.signals.EBSD(dummy_signal.data)
+        s_dict.axes_manager[0].name = "x"
+        s_dict.axes_manager[1].name = "y"
+
+        # EBSD.xmap is empty
+        with pytest.raises(ValueError, match="Dictionary signal must have a non-empty"):
+            _ = dummy_signal.dictionary_indexing(s_dict)
+
+        # EBSD not 1 navigation dimension
+        s_dict.xmap = CrystalMap.empty((3, 3))
+        with pytest.raises(ValueError, match="Dictionary signal must have a non-empty"):
+            _ = dummy_signal.dictionary_indexing(s_dict)
 
 
 class TestEBSDRefinement: